requirements.txt
numpy
PyYAML
opencv-contrib-python
